fix(cart): fix unify naming style

+++ back-nest/src/cart/cart.service.ts
@@ -9,23 +9,23 @@ export class CartService {
 
   constructor(private prisma: PrismaService) {}
 
-  async getUserCart(user_id: number) {
+  async getUserCart(userId: number) {
     const cart = await this.prisma.cart.findMany({
-      where: { user_id },
+      where: { user_id: userId }, // Using snake_case for database field
       include: { product: true },
     });
 
-    this.logger.debug(`Cart for user ${user_id}: ${JSON.stringify(cart)}`);
+    this.logger.debug(`Cart for user ${userId}: ${JSON.stringify(cart)}`);
     return cart;
   }
 
-  async updateCart({ user_id, cartItems }: UpdateCartParams) {
-    this.logger.log(`Deleting cart items for user_id=${user_id}`);
+  async updateCart({ userId, cartItems }: UpdateCartParams) {
+    this.logger.log(`Deleting cart items for userId=${userId}`);
 
-    await this.prisma.cart.deleteMany({ where: { user_id } });
+    await this.prisma.cart.deleteMany({ where: { user_id: userId } });
 
     const remainingItems = await this.prisma.cart.findMany({
-      where: { user_id },
+      where: { user_id: userId },
     });
     this.logger.debug(
       `Cart after deleteMany: ${JSON.stringify(remainingItems)}`,
@@ -46,12 +46,12 @@ export class CartService {
           data: {
             product_id: item.id,
             quantity: item.quantity,
-            user_id,
+            user_id: userId,
           },
         });
       }
     });
 
-    this.logger.log(`Successfully updated cart for user_id=${user_id}`);
+    this.logger.log(`Successfully updated cart for userId=${userId}`);
   }
 }
